From frontend-cardlist into frontend

See merge request s12-webmobile2-sub1/S12P11A805!96

+++ moda/app/src/main/java/com/example/modapjt/data/dto/response/CardDTO.kt
@@ -13,7 +13,7 @@ data class CardDTO(
     val keywords: List<String>?,
     val excludedKeywords: List<String>?,
     val isMine: Boolean,
-    val score: Int?
+    val score: Float?
 )
 
 // DTO → 도메인 모델 변환 함수
