+++ moda/app/src/main/java/com/example/modapjt/domain/model/Card.kt
@@ -13,5 +13,5 @@ data class Card(
 //    val createdAt: String,
     val excludedKeywords: List<String>?,
     val isMine: Boolean,
-    val score: Int?
+    val score: Float?
 )
